Skip unnecessary generations

+++ scripts/downsample_resolution.py
@@ -4,8 +4,8 @@ import cv2
 import numpy as np
 
 if __name__ == '__main__':
-    imgdir = os.path.expanduser(r"D:\wp-gen\livingroom\camera\temp")
-    outdir = os.path.expanduser(r"D:\wp-gen\livingroom\camera\temp_half_resolution")
+    imgdir = os.path.expanduser(r"D:\blender\scripts\BlenderCV\gttanabata_test1\raw")
+    outdir = os.path.expanduser(r"D:\blender\scripts\BlenderCV\gttanabata_test1\raw_half_resolution")
     os.makedirs(outdir, exist_ok=True)
     rate = 2
 
@@ -20,8 +20,10 @@ if __name__ == '__main__':
     w2 = hw[1] // rate
 
     for i, f in enumerate(imgfiles):
-        img = cv2.imread(f)
         name = os.path.splitext(os.path.basename(f))[0]
+        if os.path.exists(os.path.join(outdir, f"{name}.png")):
+            continue
+        img = cv2.imread(f)
         img_out = cv2.resize(img, (w2, h2), interpolation=cv2.INTER_LINEAR)
         cv2.imwrite(os.path.join(outdir, f"{name}.png"), img_out)
